mover SeedDatabase para a pasta configs
